add tambah data dan hapus data

+++ mahasiswa.php
@@ -56,7 +56,7 @@
                 </td>
                 <td>
                     <a href="editdata.php?id=<?php echo $mhs["id"]; ?>"><button>Edit</button></a> | 
-                    <a href="hapusdata.php?id=<?php echo $mhs["id"]; ?>"><button>Hapus</button></a>
+                    <a href="hapusdata.php?id=<?php echo $mhs["id"]; ?>" onclick="return confirm('Anda Yakin')" ><button>Hapus</button></a>
                 </td> 
             </tr>
             <?php
